refactor: move createXcmTypes into handlers/ dir
Just a slight reorganization to reduce the number of files/dirs per dir
