+++ src/createXcmTypes/handlers/ParaToEthereum.spec.ts
@@ -1,14 +1,14 @@
 // Copyright 2023 Parity Technologies (UK) Ltd.
 
-import { Registry } from '../registry';
-import { mockRelayApiV9420 } from '../testHelpers/mockRelayApiV9420';
-import { RelayToSystem } from './RelayToSystem';
+import { Registry } from '../../registry';
+import { mockMoonriverParachainApi } from '../../testHelpers/mockMoonriverParachainApi';
+import { ParaToEthereum } from './ParaToEthereum';
 
-describe('RelayToSystem XcmVersioned Generation', () => {
+describe('ParaToEthereum', () => {
 	const registry = new Registry('kusama', {});
 	describe('Beneficiary', () => {
 		it('Should work for V2', () => {
-			const beneficiary = RelayToSystem.createBeneficiary(
+			const beneficiary = ParaToEthereum.createBeneficiary(
 				'0xf5d5714c084c112843aca74f8c498da06cc5a2d63153b825189baa51043b1f0b',
 				2,
 			);
@@ -30,7 +30,7 @@ describe('RelayToSystem XcmVersioned Generation', () => {
 			expect(beneficiary).toStrictEqual(expectedRes);
 		});
 		it('Should work for V3', () => {
-			const beneficiary = RelayToSystem.createBeneficiary(
+			const beneficiary = ParaToEthereum.createBeneficiary(
 				'0xf5d5714c084c112843aca74f8c498da06cc5a2d63153b825189baa51043b1f0b',
 				3,
 			);
@@ -51,7 +51,7 @@ describe('RelayToSystem XcmVersioned Generation', () => {
 			expect(beneficiary).toStrictEqual(expectedRes);
 		});
 		it('Should work for V4', () => {
-			const beneficiary = RelayToSystem.createBeneficiary(
+			const beneficiary = ParaToEthereum.createBeneficiary(
 				'0xf5d5714c084c112843aca74f8c498da06cc5a2d63153b825189baa51043b1f0b',
 				4,
 			);
@@ -74,7 +74,7 @@ describe('RelayToSystem XcmVersioned Generation', () => {
 			expect(beneficiary).toStrictEqual(expectedRes);
 		});
 		it('Should work for V5', () => {
-			const beneficiary = RelayToSystem.createBeneficiary(
+			const beneficiary = ParaToEthereum.createBeneficiary(
 				'0xf5d5714c084c112843aca74f8c498da06cc5a2d63153b825189baa51043b1f0b',
 				5,
 			);
@@ -97,14 +97,13 @@ describe('RelayToSystem XcmVersioned Generation', () => {
 			expect(beneficiary).toStrictEqual(expectedRes);
 		});
 	});
-
 	describe('Destination', () => {
 		it('Should work for V2', () => {
-			const destination = RelayToSystem.createDest('100', 2);
+			const destination = ParaToEthereum.createDest('100', 2);
 
 			const expectedRes = {
 				V2: {
-					parents: 0,
+					parents: 1,
 					interior: {
 						X1: {
 							Parachain: '100',
@@ -116,11 +115,11 @@ describe('RelayToSystem XcmVersioned Generation', () => {
 			expect(destination).toStrictEqual(expectedRes);
 		});
 		it('Should work for V3', () => {
-			const destination = RelayToSystem.createDest('100', 3);
+			const destination = ParaToEthereum.createDest('100', 3);
 
 			const expectedRes = {
 				V3: {
-					parents: 0,
+					parents: 1,
 					interior: {
 						X1: {
 							Parachain: '100',
@@ -132,11 +131,11 @@ describe('RelayToSystem XcmVersioned Generation', () => {
 			expect(destination).toStrictEqual(expectedRes);
 		});
 		it('Should work for V4', () => {
-			const destination = RelayToSystem.createDest('100', 4);
+			const destination = ParaToEthereum.createDest('100', 4);
 
 			const expectedRes = {
 				V4: {
-					parents: 0,
+					parents: 1,
 					interior: {
 						X1: [
 							{
@@ -150,11 +149,11 @@ describe('RelayToSystem XcmVersioned Generation', () => {
 			expect(destination).toStrictEqual(expectedRes);
 		});
 		it('Should work for V5', () => {
-			const destination = RelayToSystem.createDest('100', 5);
+			const destination = ParaToEthereum.createDest('100', 5);
 
 			const expectedRes = {
 				V5: {
-					parents: 0,
+					parents: 1,
 					interior: {
 						X1: [
 							{
@@ -169,29 +168,48 @@ describe('RelayToSystem XcmVersioned Generation', () => {
 		});
 	});
 	describe('Assets', () => {
-		const isForeignAssetsTransfer = false;
 		const isLiquidTokenTransfer = false;
+		const isForeignAssetsTransfer = false;
 		it('Should work for V2', async () => {
-			const assets = await RelayToSystem.createAssets(['100'], 2, '', [], {
-				registry,
-				isForeignAssetsTransfer,
-				isLiquidTokenTransfer,
-				api: mockRelayApiV9420,
-			});
+			const assets = await ParaToEthereum.createAssets(
+				['1000000000000', '2000000000'],
+				2,
+				'moonriver',
+				['42259045809535163221576417993425387648', '182365888117048807484804376330534607370'],
+				{
+					registry,
+					isForeignAssetsTransfer,
+					isLiquidTokenTransfer,
+					api: mockMoonriverParachainApi,
+				},
+			);
 
 			const expectedRes = {
 				V2: [
 					{
-						fun: {
-							Fungible: '100',
-						},
 						id: {
 							Concrete: {
-								interior: {
-									Here: '',
+								Parents: '1',
+								Interior: {
+									Here: null,
 								},
-								parents: 0,
 							},
+						},
+						fun: {
+							Fungible: '1000000000000',
+						},
+					},
+					{
+						id: {
+							Concrete: {
+								Parents: '1',
+								Interior: {
+									X3: [{ Parachain: '1000' }, { PalletInstance: '50' }, { GeneralIndex: '8' }],
+								},
+							},
+						},
+						fun: {
+							Fungible: '2000000000',
 						},
 					},
 				],
@@ -200,26 +218,45 @@ describe('RelayToSystem XcmVersioned Generation', () => {
 			expect(assets).toStrictEqual(expectedRes);
 		});
 		it('Should work for V3', async () => {
-			const assets = await RelayToSystem.createAssets(['100'], 3, '', [], {
-				registry,
-				isForeignAssetsTransfer,
-				isLiquidTokenTransfer,
-				api: mockRelayApiV9420,
-			});
+			const assets = await ParaToEthereum.createAssets(
+				['1000000', '20000000000'],
+				3,
+				'moonriver',
+				['182365888117048807484804376330534607370', '311091173110107856861649819128533077277'],
+				{
+					registry,
+					isForeignAssetsTransfer,
+					isLiquidTokenTransfer,
+					api: mockMoonriverParachainApi,
+				},
+			);
 
 			const expectedRes = {
 				V3: [
 					{
-						fun: {
-							Fungible: '100',
-						},
 						id: {
 							Concrete: {
-								interior: {
-									Here: '',
+								Parents: '1',
+								Interior: {
+									X3: [{ Parachain: '1000' }, { PalletInstance: '50' }, { GeneralIndex: '8' }],
 								},
-								parents: 0,
 							},
+						},
+						fun: {
+							Fungible: '1000000',
+						},
+					},
+					{
+						id: {
+							Concrete: {
+								Parents: '1',
+								Interior: {
+									X3: [{ Parachain: '1000' }, { PalletInstance: '50' }, { GeneralIndex: '1984' }],
+								},
+							},
+						},
+						fun: {
+							Fungible: '20000000000',
 						},
 					},
 				],
@@ -228,24 +265,41 @@ describe('RelayToSystem XcmVersioned Generation', () => {
 			expect(assets).toStrictEqual(expectedRes);
 		});
 		it('Should work for V4', async () => {
-			const assets = await RelayToSystem.createAssets(['100'], 4, '', [], {
-				registry,
-				isForeignAssetsTransfer,
-				isLiquidTokenTransfer,
-				api: mockRelayApiV9420,
-			});
+			const assets = await ParaToEthereum.createAssets(
+				['1000000', '20000000000'],
+				4,
+				'moonriver',
+				['182365888117048807484804376330534607370', '311091173110107856861649819128533077277'],
+				{
+					registry,
+					isForeignAssetsTransfer,
+					isLiquidTokenTransfer,
+					api: mockMoonriverParachainApi,
+				},
+			);
 
 			const expectedRes = {
 				V4: [
 					{
-						fun: {
-							Fungible: '100',
-						},
 						id: {
-							interior: {
-								Here: '',
+							Parents: '1',
+							Interior: {
+								X3: [{ Parachain: '1000' }, { PalletInstance: '50' }, { GeneralIndex: '8' }],
 							},
-							parents: 0,
+						},
+						fun: {
+							Fungible: '1000000',
+						},
+					},
+					{
+						id: {
+							Parents: '1',
+							Interior: {
+								X3: [{ Parachain: '1000' }, { PalletInstance: '50' }, { GeneralIndex: '1984' }],
+							},
+						},
+						fun: {
+							Fungible: '20000000000',
 						},
 					},
 				],
@@ -254,24 +308,41 @@ describe('RelayToSystem XcmVersioned Generation', () => {
 			expect(assets).toStrictEqual(expectedRes);
 		});
 		it('Should work for V5', async () => {
-			const assets = await RelayToSystem.createAssets(['100'], 5, '', [], {
-				registry,
-				isForeignAssetsTransfer,
-				isLiquidTokenTransfer,
-				api: mockRelayApiV9420,
-			});
+			const assets = await ParaToEthereum.createAssets(
+				['1000000', '20000000000'],
+				5,
+				'moonriver',
+				['182365888117048807484804376330534607370', '311091173110107856861649819128533077277'],
+				{
+					registry,
+					isForeignAssetsTransfer,
+					isLiquidTokenTransfer,
+					api: mockMoonriverParachainApi,
+				},
+			);
 
 			const expectedRes = {
 				V5: [
 					{
-						fun: {
-							Fungible: '100',
-						},
 						id: {
-							interior: {
-								Here: '',
+							Parents: '1',
+							Interior: {
+								X3: [{ Parachain: '1000' }, { PalletInstance: '50' }, { GeneralIndex: '8' }],
 							},
-							parents: 0,
+						},
+						fun: {
+							Fungible: '1000000',
+						},
+					},
+					{
+						id: {
+							Parents: '1',
+							Interior: {
+								X3: [{ Parachain: '1000' }, { PalletInstance: '50' }, { GeneralIndex: '1984' }],
+							},
+						},
+						fun: {
+							Fungible: '20000000000',
 						},
 					},
 				],
@@ -281,13 +352,11 @@ describe('RelayToSystem XcmVersioned Generation', () => {
 		});
 	});
 	describe('WeightLimit', () => {
-		// NOTE: for V0, V1, and V2 Weightlimit just uses V2 so we only need to test once.
-		// No matter the version if its equal to or less than 2, it will alwyas default to V2.
 		it('Should work when weightLimit option is provided', () => {
 			const refTime = '100000000';
 			const proofSize = '1000';
 
-			const weightLimit = RelayToSystem.createWeightLimit({
+			const weightLimit = ParaToEthereum.createWeightLimit({
 				weightLimit: {
 					refTime,
 					proofSize,
@@ -295,13 +364,13 @@ describe('RelayToSystem XcmVersioned Generation', () => {
 			});
 			expect(weightLimit).toStrictEqual({
 				Limited: {
-					refTime: '100000000',
 					proofSize: '1000',
+					refTime: '100000000',
 				},
 			});
 		});
 		it('Should work when weightLimit option is not provided', () => {
-			const weightLimit = RelayToSystem.createWeightLimit({});
+			const weightLimit = ParaToEthereum.createWeightLimit({});
 
 			expect(weightLimit).toStrictEqual({
 				Unlimited: null,
