+++ src/createXcmTypes/handlers/ParaToEthereum.ts
@@ -3,36 +3,33 @@
 import type { ApiPromise } from '@polkadot/api';
 import type { AnyJson } from '@polkadot/types/types';
 
-import { DEFAULT_XCM_VERSION } from '../consts.js';
-import { Registry } from '../registry/index.js';
-import { XCMAssetRegistryMultiLocation } from '../registry/types.js';
-import { Direction } from '../types.js';
-import { resolveMultiLocation } from '../util/resolveMultiLocation.js';
+import { DEFAULT_XCM_VERSION } from '../../consts.js';
+import { Registry } from '../../registry/index.js';
+import { XCMAssetRegistryMultiLocation } from '../../registry/types.js';
+import { Direction } from '../../types.js';
+import { resolveMultiLocation } from '../../util/resolveMultiLocation.js';
 import type {
 	CreateAssetsOpts,
 	CreateFeeAssetItemOpts,
 	FungibleAssetType,
+	FungibleMultiAsset,
 	ICreateXcmType,
-	UnionXcAssetsMultiAsset,
-	UnionXcAssetsMultiAssets,
 	UnionXcmMultiAssets,
 	XcmDestBeneficiary,
-} from './types.js';
-import { createAssets } from './util/createAssets.js';
-import { createBeneficiary, createXTokensParachainDestBeneficiary } from './util/createBeneficiary.js';
-import { createParachainDest } from './util/createDest.js';
-import { createFeeAssetItem } from './util/createFeeAssetItem.js';
-import { createMultiAsset } from './util/createMultiAsset.js';
-import { createWeightLimit } from './util/createWeightLimit.js';
-import { createXTokensAsset, createXTokensMultiAssets } from './util/createXTokensAssets.js';
-import { createXTokensFeeAssetItem } from './util/createXTokensFeeAssetItem.js';
-import { dedupeAssets } from './util/dedupeAssets.js';
-import { getParachainNativeAssetLocation } from './util/getParachainNativeAssetLocation.js';
-import { getXcAssetMultiLocationByAssetId } from './util/getXcAssetMultiLocationByAssetId.js';
-import { isParachainPrimaryNativeAsset } from './util/isParachainPrimaryNativeAsset.js';
-import { sortAssetsAscending } from './util/sortAssetsAscending.js';
+} from '../types.js';
+import { createAssets } from '../util/createAssets.js';
+import { createBeneficiary } from '../util/createBeneficiary.js';
+import { createParachainDest } from '../util/createDest.js';
+import { createFeeAssetItem } from '../util/createFeeAssetItem.js';
+import { createMultiAsset } from '../util/createMultiAsset.js';
+import { createWeightLimit } from '../util/createWeightLimit.js';
+import { dedupeAssets } from '../util/dedupeAssets.js';
+import { getParachainNativeAssetLocation } from '../util/getParachainNativeAssetLocation.js';
+import { getXcAssetMultiLocationByAssetId } from '../util/getXcAssetMultiLocationByAssetId.js';
+import { isParachainPrimaryNativeAsset } from '../util/isParachainPrimaryNativeAsset.js';
+import { sortAssetsAscending } from '../util/sortAssetsAscending.js';
 
-export const ParaToSystem: ICreateXcmType = {
+export const ParaToEthereum: ICreateXcmType = {
 	/**
 	 * Create a XcmVersionedMultiLocation type for a beneficiary.
 	 *
@@ -75,7 +72,7 @@ export const ParaToSystem: ICreateXcmType = {
 			specName,
 			assets,
 			opts,
-			multiAssetCreator: createParaToSystemMultiAssets,
+			multiAssetCreator: createParaToEthereumMultiAssets,
 		});
 	},
 	/**
@@ -94,74 +91,13 @@ export const ParaToSystem: ICreateXcmType = {
 		return createFeeAssetItem({
 			api,
 			opts,
-			multiAssetCreator: createParaToSystemMultiAssets,
+			multiAssetCreator: createParaToEthereumMultiAssets,
 		});
 	},
-	/**
-	 * Create xTokens beneficiary structured type.
-	 *
-	 * @param destChainId The parachain Id of the destination.
-	 * @param accountId The accountId of the beneficiary.
-	 * @param xcmVersion The accepted xcm version.
-	 */
-	createXTokensBeneficiary: createXTokensParachainDestBeneficiary,
-	/**
-	 * Create multiple xTokens Assets.
-	 *
-	 * @param amounts Amount per asset. It will match the `assets` length.
-	 * @param xcmVersion The accepted xcm version.
-	 * @param specName The specname of the chain the api is connected to.
-	 * @param assets The assets to create into xcm `MultiAssets`.
-	 * @param opts Options used to create xTokens `MultiAssets`.
-	 */
-	createXTokensAssets: async (
-		amounts: string[],
-		xcmVersion: number,
-		specName: string,
-		assets: string[],
-		opts: CreateAssetsOpts,
-	): Promise<UnionXcAssetsMultiAssets> => {
-		return createXTokensMultiAssets({
-			amounts,
-			assets,
-			opts,
-			specName,
-			xcmVersion,
-		});
-	},
-	/**
-	 * Create a single xToken asset.
-	 *
-	 * @param amount Amount per asset. This will be of length 1.
-	 * @param xcmVersion The accepted xcm version.
-	 * @param specName The specname of the chain the api is connected to.
-	 * @param assetId Single asset to be created into a `MultiAsset`.
-	 * @param opts Options to create a single Asset.
-	 */
-	createXTokensAsset: async (
-		amount: string,
-		xcmVersion: number,
-		specName: string,
-		assetId: string,
-		opts: CreateAssetsOpts,
-	): Promise<UnionXcAssetsMultiAsset> => {
-		return createXTokensAsset({
-			amount,
-			assetId,
-			opts,
-			specName,
-			xcmVersion,
-		});
-	},
-	/**
-	 * Create an xTokens xcm `feeAssetItem`.
-	 *
-	 * @param opts Options used for creating `feeAssetItem`.
-	 */
-	createXTokensFeeAssetItem,
 };
+
 /**
- * Create multiassets for ParaToSystem direction.
+ * Create multiassets for ParaToEthereum direction.
  *
  * @param api ApiPromise
  * @param amounts Amount per asset. It will match the `assets` length.
@@ -171,7 +107,7 @@ export const ParaToSystem: ICreateXcmType = {
  * @param registry The asset registry used to construct MultiLocations.
  * @param isForeignAssetsTransfer Whether this transfer is a foreign assets transfer.
  */
-const createParaToSystemMultiAssets = async ({
+const createParaToEthereumMultiAssets = async ({
 	api,
 	amounts,
 	specName,
@@ -188,11 +124,11 @@ const createParaToSystemMultiAssets = async ({
 	registry: Registry;
 	destChainId?: string;
 }): Promise<FungibleAssetType[]> => {
-	let multiAssets: FungibleAssetType[] = [];
+	const multiAssets: FungibleAssetType[] = [];
 	const isPrimaryParachainNativeAsset = isParachainPrimaryNativeAsset(
 		registry,
 		specName,
-		Direction.ParaToSystem,
+		Direction.ParaToEthereum,
 		assets[0],
 	);
 
@@ -225,18 +161,18 @@ const createParaToSystemMultiAssets = async ({
 			const xcAssetMultiLocation = parsedMultiLocation.v1 as unknown as AnyJson;
 
 			const multiLocation = resolveMultiLocation(xcAssetMultiLocation, xcmVersion);
-
 			const multiAsset = createMultiAsset({
-				amount: amount,
+				amount,
 				multiLocation,
 				xcmVersion,
 			});
+
 			multiAssets.push(multiAsset);
 		}
 	}
 
-	multiAssets = sortAssetsAscending(multiAssets);
-	const sortedAndDedupedMultiAssets = dedupeAssets(multiAssets);
+	const sortedAssets = sortAssetsAscending(multiAssets);
+	const sortedAndDedupedMultiAssets = dedupeAssets(sortedAssets) as FungibleMultiAsset[];
 
 	return sortedAndDedupedMultiAssets;
 };
