+++ src/createXcmTypes/handlers/ParaToRelay.spec.ts
@@ -1,7 +1,7 @@
 // Copyright 2023 Parity Technologies (UK) Ltd.
 
-import { Registry } from '../registry';
-import { adjustedMockMoonriverParachainApi } from '../testHelpers/adjustedMockMoonriverParachainApi';
+import { Registry } from '../../registry';
+import { adjustedMockMoonriverParachainApi } from '../../testHelpers/adjustedMockMoonriverParachainApi';
 import { ParaToRelay } from './ParaToRelay';
 
 describe('ParaToRelay', () => {
