+++ src/createXcmTypes/handlers/RelayToBridge.spec.ts
@@ -1,16 +1,16 @@
 // Copyright 2024 Parity Technologies (UK) Ltd.
 
-import { Registry } from '../registry';
-import { adjustedMockSystemApiV1016000 } from '../testHelpers/adjustedMockSystemApiV1016000';
-import { SystemToBridge } from './SystemToBridge';
+import { Registry } from '../../registry';
+import { adjustedMockSystemApiV1016000 } from '../../testHelpers/adjustedMockSystemApiV1016000';
+import { RelayToBridge } from './RelayToBridge';
 
-describe('SystemToBridge', () => {
-	const registry = new Registry('asset-hub-paseo', {});
+describe('RelayToBridge', () => {
+	const registry = new Registry('paseo', {});
 	const isForeignAssetsTransfer = true;
 	const isLiquidTokenTransfer = false;
 	describe('Beneficiary', () => {
 		it('Should work for V3', () => {
-			const beneficiary = SystemToBridge.createBeneficiary(
+			const beneficiary = RelayToBridge.createBeneficiary(
 				'0xf5d5714c084c112843aca74f8c498da06cc5a2d63153b825189baa51043b1f0b',
 				3,
 			);
@@ -31,7 +31,7 @@ describe('SystemToBridge', () => {
 			expect(beneficiary).toStrictEqual(expectedRes);
 		});
 		it('Should work for V3 for an Ethereum Address', () => {
-			const beneficiary = SystemToBridge.createBeneficiary('0x96Bd611EbE3Af39544104e26764F4939924F6Ece', 3);
+			const beneficiary = RelayToBridge.createBeneficiary('0x96Bd611EbE3Af39544104e26764F4939924F6Ece', 3);
 
 			const expectedRes = {
 				V3: {
@@ -49,7 +49,7 @@ describe('SystemToBridge', () => {
 			expect(beneficiary).toStrictEqual(expectedRes);
 		});
 		it('Should work for V4', () => {
-			const beneficiary = SystemToBridge.createBeneficiary(
+			const beneficiary = RelayToBridge.createBeneficiary(
 				'0xf5d5714c084c112843aca74f8c498da06cc5a2d63153b825189baa51043b1f0b',
 				4,
 			);
@@ -72,7 +72,7 @@ describe('SystemToBridge', () => {
 			expect(beneficiary).toStrictEqual(expectedRes);
 		});
 		it('Should work for V5', () => {
-			const beneficiary = SystemToBridge.createBeneficiary(
+			const beneficiary = RelayToBridge.createBeneficiary(
 				'0xf5d5714c084c112843aca74f8c498da06cc5a2d63153b825189baa51043b1f0b',
 				5,
 			);
@@ -95,7 +95,7 @@ describe('SystemToBridge', () => {
 			expect(beneficiary).toStrictEqual(expectedRes);
 		});
 		it('Should work for V4 for an Ethereum Address', () => {
-			const beneficiary = SystemToBridge.createBeneficiary('0x96Bd611EbE3Af39544104e26764F4939924F6Ece', 4);
+			const beneficiary = RelayToBridge.createBeneficiary('0x96Bd611EbE3Af39544104e26764F4939924F6Ece', 4);
 
 			const expectedRes = {
 				V4: {
@@ -115,7 +115,7 @@ describe('SystemToBridge', () => {
 			expect(beneficiary).toStrictEqual(expectedRes);
 		});
 		it('Should work for V5 for an Ethereum Address', () => {
-			const beneficiary = SystemToBridge.createBeneficiary('0x96Bd611EbE3Af39544104e26764F4939924F6Ece', 5);
+			const beneficiary = RelayToBridge.createBeneficiary('0x96Bd611EbE3Af39544104e26764F4939924F6Ece', 5);
 
 			const expectedRes = {
 				V5: {
@@ -138,11 +138,11 @@ describe('SystemToBridge', () => {
 	describe('Destination', () => {
 		it('Should work for V3', () => {
 			const destId = `{"parents":"2","interior":{"X1":{"GlobalConsensus":{"Ethereum":{"chainId":"11155111"}}}}}`;
-			const destination = SystemToBridge.createDest(destId, 3);
+			const destination = RelayToBridge.createDest(destId, 3);
 
 			const expectedRes = {
 				V3: {
-					parents: 2,
+					parents: 1,
 					interior: {
 						X1: {
 							GlobalConsensus: {
@@ -160,11 +160,11 @@ describe('SystemToBridge', () => {
 		it('Should work for V4', () => {
 			const destId = `{"parents":"2","interior":{"X2":[{"GlobalConsensus":"Kusama"},{"Parachain":"1000"}]}}`;
 
-			const destination = SystemToBridge.createDest(destId, 4);
+			const destination = RelayToBridge.createDest(destId, 4);
 
 			const expectedRes = {
 				V4: {
-					parents: 2,
+					parents: 1,
 					interior: {
 						X2: [
 							{
@@ -183,11 +183,11 @@ describe('SystemToBridge', () => {
 		it('Should work for V5', () => {
 			const destId = `{"parents":"2","interior":{"X2":[{"GlobalConsensus":"Kusama"},{"Parachain":"1000"}]}}`;
 
-			const destination = SystemToBridge.createDest(destId, 5);
+			const destination = RelayToBridge.createDest(destId, 5);
 
 			const expectedRes = {
 				V5: {
-					parents: 2,
+					parents: 1,
 					interior: {
 						X2: [
 							{
@@ -206,11 +206,11 @@ describe('SystemToBridge', () => {
 	});
 	describe('Assets', () => {
 		it('Should work for V3', async () => {
-			const assets = await SystemToBridge.createAssets(
+			const assets = await RelayToBridge.createAssets(
 				['10000000000'],
 				3,
-				'asset-hub-westend',
-				[`{"parents":"2","interior":{"X1":{"GlobalConsensus":"Paseo"}}}`],
+				'paseo',
+				[`{"parents":"0","interior":{"Here":""}}`],
 				{
 					registry,
 					isForeignAssetsTransfer,
@@ -224,9 +224,9 @@ describe('SystemToBridge', () => {
 					{
 						id: {
 							Concrete: {
-								parents: '2',
+								parents: 0,
 								interior: {
-									X1: { GlobalConsensus: 'Paseo' },
+									Here: '',
 								},
 							},
 						},
@@ -240,13 +240,11 @@ describe('SystemToBridge', () => {
 			expect(assets).toStrictEqual(expectedRes);
 		});
 		it('Should work for V4', async () => {
-			const assets = await SystemToBridge.createAssets(
+			const assets = await RelayToBridge.createAssets(
 				['10000000000'],
 				4,
-				'asset-hub-westend',
-				[
-					`{"parents":"2","interior":{"X2":[{"GlobalConsensus":{"Ethereum":{"chainId":"11155111"}}},{"AccountKey20":{"network":null,"key":"0xfff9976782d46cc05630d1f6ebab18b2324d6b14"}}]}}`,
-				],
+				'paseo',
+				[`{"parents":"0","interior":{"Here":""}}`],
 				{
 					registry,
 					isForeignAssetsTransfer,
@@ -259,23 +257,9 @@ describe('SystemToBridge', () => {
 				V4: [
 					{
 						id: {
-							parents: '2',
+							parents: 0,
 							interior: {
-								X2: [
-									{
-										GlobalConsensus: {
-											Ethereum: {
-												chainId: '11155111',
-											},
-										},
-									},
-									{
-										AccountKey20: {
-											network: null,
-											key: '0xfff9976782d46cc05630d1f6ebab18b2324d6b14',
-										},
-									},
-								],
+								Here: '',
 							},
 						},
 						fun: {
@@ -288,13 +272,11 @@ describe('SystemToBridge', () => {
 			expect(assets).toStrictEqual(expectedRes);
 		});
 		it('Should work for V5', async () => {
-			const assets = await SystemToBridge.createAssets(
+			const assets = await RelayToBridge.createAssets(
 				['10000000000'],
 				5,
-				'asset-hub-westend',
-				[
-					`{"parents":"2","interior":{"X2":[{"GlobalConsensus":{"Ethereum":{"chainId":"11155111"}}},{"AccountKey20":{"network":null,"key":"0xfff9976782d46cc05630d1f6ebab18b2324d6b14"}}]}}`,
-				],
+				'paseo',
+				[`{"parents":"0","interior":{"Here":""}}`],
 				{
 					registry,
 					isForeignAssetsTransfer,
@@ -307,23 +289,9 @@ describe('SystemToBridge', () => {
 				V5: [
 					{
 						id: {
-							parents: '2',
+							parents: 0,
 							interior: {
-								X2: [
-									{
-										GlobalConsensus: {
-											Ethereum: {
-												chainId: '11155111',
-											},
-										},
-									},
-									{
-										AccountKey20: {
-											network: null,
-											key: '0xfff9976782d46cc05630d1f6ebab18b2324d6b14',
-										},
-									},
-								],
+								Here: '',
 							},
 						},
 						fun: {
