+++ src/createXcmTypes/handlers/RelayToBridge.ts
@@ -1,0 +1,59 @@
+// Copyright 2024 Parity Technologies (UK) Ltd.
+
+import type { ApiPromise } from '@polkadot/api';
+
+import { ICreateXcmType, UnionXcmMultiAssets, XcmDestBeneficiary } from '../types.js';
+import { createSingleAsset } from '../util/createAssets.js';
+import { createBeneficiary } from '../util/createBeneficiary.js';
+import { createInteriorValueDest } from '../util/createDest.js';
+import { createWeightLimit } from '../util/createWeightLimit.js';
+
+export const RelayToBridge: ICreateXcmType = {
+	/**
+	 * Create a XcmVersionedMultiLocation structured type for a beneficiary.
+	 *
+	 * @param accountId The accountId of the beneficiary.
+	 * @param xcmVersion The accepted xcm version.
+	 */
+	createBeneficiary,
+	/**
+	 * Create a XcmVersionedMultiLocation structured type for a destination.
+	 *
+	 * @param destId The chainId of the destination.
+	 * @param xcmVersion The accepted xcm version.
+	 */
+	createDest: (destId: string, xcmVersion: number): XcmDestBeneficiary => {
+		return createInteriorValueDest({
+			destId,
+			parents: 1,
+			xcmVersion,
+		});
+	},
+	/**
+	 * Create a VersionedMultiAsset structured type.
+	 *
+	 * @param amounts Amount per asset. It will match the `assets` length.
+	 * @param xcmVersion The accepted xcm version.
+	 */
+	createAssets: async (amounts: string[], xcmVersion: number): Promise<UnionXcmMultiAssets> => {
+		return createSingleAsset({
+			amounts,
+			parents: 0,
+			xcmVersion,
+		});
+	},
+	/**
+	 * Create an Xcm WeightLimit structured type.
+	 *
+	 * @param opts Options that are used for WeightLimit.
+	 */
+	createWeightLimit,
+	/**
+	 * Returns the correct `feeAssetItem` based on XCM direction.
+	 *
+	 * @param api ApiPromise
+	 */
+	createFeeAssetItem: async (_: ApiPromise): Promise<number> => {
+		return Promise.resolve(0);
+	},
+};
