+++ src/createXcmTypes/handlers/RelayToPara.spec.ts
@@ -1,14 +1,14 @@
 // Copyright 2023 Parity Technologies (UK) Ltd.
 
-import { Registry } from '../registry';
-import { mockMoonriverParachainApi } from '../testHelpers/mockMoonriverParachainApi';
-import { ParaToSystem } from './ParaToSystem';
+import { Registry } from '../../registry';
+import { mockRelayApiV9420 } from '../../testHelpers/mockRelayApiV9420';
+import { RelayToPara } from './RelayToPara';
 
-describe('ParaToSystem', () => {
+describe('RelayToPara XcmVersioned Generation', () => {
 	const registry = new Registry('kusama', {});
 	describe('Beneficiary', () => {
 		it('Should work for V2', () => {
-			const beneficiary = ParaToSystem.createBeneficiary(
+			const beneficiary = RelayToPara.createBeneficiary(
 				'0xf5d5714c084c112843aca74f8c498da06cc5a2d63153b825189baa51043b1f0b',
 				2,
 			);
@@ -29,8 +29,27 @@ describe('ParaToSystem', () => {
 
 			expect(beneficiary).toStrictEqual(expectedRes);
 		});
+		it('Should work for V2 Ethereum address', () => {
+			const beneficiary = RelayToPara.createBeneficiary('0x96Bd611EbE3Af39544104e26764F4939924F6Ece', 2);
+
+			const expectedRes = {
+				V2: {
+					parents: 0,
+					interior: {
+						X1: {
+							AccountKey20: {
+								key: '0x96Bd611EbE3Af39544104e26764F4939924F6Ece',
+								network: 'Any',
+							},
+						},
+					},
+				},
+			};
+
+			expect(beneficiary).toStrictEqual(expectedRes);
+		});
 		it('Should work for V3', () => {
-			const beneficiary = ParaToSystem.createBeneficiary(
+			const beneficiary = RelayToPara.createBeneficiary(
 				'0xf5d5714c084c112843aca74f8c498da06cc5a2d63153b825189baa51043b1f0b',
 				3,
 			);
@@ -50,8 +69,27 @@ describe('ParaToSystem', () => {
 
 			expect(beneficiary).toStrictEqual(expectedRes);
 		});
+		it('Should work for V3 Ethereum address', () => {
+			const beneficiary = RelayToPara.createBeneficiary('0x96Bd611EbE3Af39544104e26764F4939924F6Ece', 3);
+
+			const expectedRes = {
+				V3: {
+					parents: 0,
+					interior: {
+						X1: {
+							AccountKey20: {
+								key: '0x96Bd611EbE3Af39544104e26764F4939924F6Ece',
+							},
+						},
+					},
+				},
+			};
+
+			expect(beneficiary).toStrictEqual(expectedRes);
+		});
+
 		it('Should work for V4', () => {
-			const beneficiary = ParaToSystem.createBeneficiary(
+			const beneficiary = RelayToPara.createBeneficiary(
 				'0xf5d5714c084c112843aca74f8c498da06cc5a2d63153b825189baa51043b1f0b',
 				4,
 			);
@@ -74,7 +112,7 @@ describe('ParaToSystem', () => {
 			expect(beneficiary).toStrictEqual(expectedRes);
 		});
 		it('Should work for V5', () => {
-			const beneficiary = ParaToSystem.createBeneficiary(
+			const beneficiary = RelayToPara.createBeneficiary(
 				'0xf5d5714c084c112843aca74f8c498da06cc5a2d63153b825189baa51043b1f0b',
 				5,
 			);
@@ -96,14 +134,55 @@ describe('ParaToSystem', () => {
 
 			expect(beneficiary).toStrictEqual(expectedRes);
 		});
+		it('Should work for V4 Ethereum address', () => {
+			const beneficiary = RelayToPara.createBeneficiary('0x96Bd611EbE3Af39544104e26764F4939924F6Ece', 4);
+
+			const expectedRes = {
+				V4: {
+					parents: 0,
+					interior: {
+						X1: [
+							{
+								AccountKey20: {
+									key: '0x96Bd611EbE3Af39544104e26764F4939924F6Ece',
+								},
+							},
+						],
+					},
+				},
+			};
+
+			expect(beneficiary).toStrictEqual(expectedRes);
+		});
+		it('Should work for V5 Ethereum address', () => {
+			const beneficiary = RelayToPara.createBeneficiary('0x96Bd611EbE3Af39544104e26764F4939924F6Ece', 5);
+
+			const expectedRes = {
+				V5: {
+					parents: 0,
+					interior: {
+						X1: [
+							{
+								AccountKey20: {
+									key: '0x96Bd611EbE3Af39544104e26764F4939924F6Ece',
+								},
+							},
+						],
+					},
+				},
+			};
+
+			expect(beneficiary).toStrictEqual(expectedRes);
+		});
 	});
+
 	describe('Destination', () => {
 		it('Should work for V2', () => {
-			const destination = ParaToSystem.createDest('100', 2);
+			const destination = RelayToPara.createDest('100', 2);
 
 			const expectedRes = {
 				V2: {
-					parents: 1,
+					parents: 0,
 					interior: {
 						X1: {
 							Parachain: '100',
@@ -115,11 +194,11 @@ describe('ParaToSystem', () => {
 			expect(destination).toStrictEqual(expectedRes);
 		});
 		it('Should work for V3', () => {
-			const destination = ParaToSystem.createDest('100', 3);
+			const destination = RelayToPara.createDest('100', 3);
 
 			const expectedRes = {
 				V3: {
-					parents: 1,
+					parents: 0,
 					interior: {
 						X1: {
 							Parachain: '100',
@@ -131,11 +210,11 @@ describe('ParaToSystem', () => {
 			expect(destination).toStrictEqual(expectedRes);
 		});
 		it('Should work for V4', () => {
-			const destination = ParaToSystem.createDest('100', 4);
+			const destination = RelayToPara.createDest('100', 4);
 
 			const expectedRes = {
 				V4: {
-					parents: 1,
+					parents: 0,
 					interior: {
 						X1: [
 							{
@@ -149,11 +228,11 @@ describe('ParaToSystem', () => {
 			expect(destination).toStrictEqual(expectedRes);
 		});
 		it('Should work for V5', () => {
-			const destination = ParaToSystem.createDest('100', 5);
+			const destination = RelayToPara.createDest('100', 5);
 
 			const expectedRes = {
 				V5: {
-					parents: 1,
+					parents: 0,
 					interior: {
 						X1: [
 							{
@@ -168,48 +247,30 @@ describe('ParaToSystem', () => {
 		});
 	});
 	describe('Assets', () => {
-		const isLiquidTokenTransfer = false;
 		const isForeignAssetsTransfer = false;
+		const isLiquidTokenTransfer = false;
+
 		it('Should work for V2', async () => {
-			const assets = await ParaToSystem.createAssets(
-				['1000000000000', '2000000000'],
-				2,
-				'moonriver',
-				['42259045809535163221576417993425387648', '182365888117048807484804376330534607370'],
-				{
-					registry,
-					isForeignAssetsTransfer,
-					isLiquidTokenTransfer,
-					api: mockMoonriverParachainApi,
-				},
-			);
+			const assets = await RelayToPara.createAssets(['100'], 2, '', [], {
+				registry,
+				isForeignAssetsTransfer,
+				isLiquidTokenTransfer,
+				api: mockRelayApiV9420,
+			});
 
 			const expectedRes = {
 				V2: [
 					{
+						fun: {
+							Fungible: '100',
+						},
 						id: {
 							Concrete: {
-								Parents: '1',
-								Interior: {
-									Here: null,
+								interior: {
+									Here: '',
 								},
+								parents: 0,
 							},
-						},
-						fun: {
-							Fungible: '1000000000000',
-						},
-					},
-					{
-						id: {
-							Concrete: {
-								Parents: '1',
-								Interior: {
-									X3: [{ Parachain: '1000' }, { PalletInstance: '50' }, { GeneralIndex: '8' }],
-								},
-							},
-						},
-						fun: {
-							Fungible: '2000000000',
 						},
 					},
 				],
@@ -218,45 +279,26 @@ describe('ParaToSystem', () => {
 			expect(assets).toStrictEqual(expectedRes);
 		});
 		it('Should work for V3', async () => {
-			const assets = await ParaToSystem.createAssets(
-				['1000000', '20000000000'],
-				3,
-				'moonriver',
-				['182365888117048807484804376330534607370', '311091173110107856861649819128533077277'],
-				{
-					registry,
-					isForeignAssetsTransfer,
-					isLiquidTokenTransfer,
-					api: mockMoonriverParachainApi,
-				},
-			);
+			const assets = await RelayToPara.createAssets(['100'], 3, '', [], {
+				registry,
+				isForeignAssetsTransfer,
+				isLiquidTokenTransfer,
+				api: mockRelayApiV9420,
+			});
 
 			const expectedRes = {
 				V3: [
 					{
+						fun: {
+							Fungible: '100',
+						},
 						id: {
 							Concrete: {
-								Parents: '1',
-								Interior: {
-									X3: [{ Parachain: '1000' }, { PalletInstance: '50' }, { GeneralIndex: '8' }],
+								interior: {
+									Here: '',
 								},
+								parents: 0,
 							},
-						},
-						fun: {
-							Fungible: '1000000',
-						},
-					},
-					{
-						id: {
-							Concrete: {
-								Parents: '1',
-								Interior: {
-									X3: [{ Parachain: '1000' }, { PalletInstance: '50' }, { GeneralIndex: '1984' }],
-								},
-							},
-						},
-						fun: {
-							Fungible: '20000000000',
 						},
 					},
 				],
@@ -265,41 +307,24 @@ describe('ParaToSystem', () => {
 			expect(assets).toStrictEqual(expectedRes);
 		});
 		it('Should work for V4', async () => {
-			const assets = await ParaToSystem.createAssets(
-				['1000000', '20000000000'],
-				4,
-				'moonriver',
-				['182365888117048807484804376330534607370', '311091173110107856861649819128533077277'],
-				{
-					registry,
-					isForeignAssetsTransfer,
-					isLiquidTokenTransfer,
-					api: mockMoonriverParachainApi,
-				},
-			);
+			const assets = await RelayToPara.createAssets(['100'], 4, '', [], {
+				registry,
+				isForeignAssetsTransfer,
+				isLiquidTokenTransfer,
+				api: mockRelayApiV9420,
+			});
 
 			const expectedRes = {
 				V4: [
 					{
-						id: {
-							Parents: '1',
-							Interior: {
-								X3: [{ Parachain: '1000' }, { PalletInstance: '50' }, { GeneralIndex: '8' }],
-							},
-						},
 						fun: {
-							Fungible: '1000000',
+							Fungible: '100',
 						},
-					},
-					{
 						id: {
-							Parents: '1',
-							Interior: {
-								X3: [{ Parachain: '1000' }, { PalletInstance: '50' }, { GeneralIndex: '1984' }],
+							interior: {
+								Here: '',
 							},
-						},
-						fun: {
-							Fungible: '20000000000',
+							parents: 0,
 						},
 					},
 				],
@@ -308,41 +333,24 @@ describe('ParaToSystem', () => {
 			expect(assets).toStrictEqual(expectedRes);
 		});
 		it('Should work for V5', async () => {
-			const assets = await ParaToSystem.createAssets(
-				['1000000', '20000000000'],
-				5,
-				'moonriver',
-				['182365888117048807484804376330534607370', '311091173110107856861649819128533077277'],
-				{
-					registry,
-					isForeignAssetsTransfer,
-					isLiquidTokenTransfer,
-					api: mockMoonriverParachainApi,
-				},
-			);
+			const assets = await RelayToPara.createAssets(['100'], 5, '', [], {
+				registry,
+				isForeignAssetsTransfer,
+				isLiquidTokenTransfer,
+				api: mockRelayApiV9420,
+			});
 
 			const expectedRes = {
 				V5: [
 					{
-						id: {
-							Parents: '1',
-							Interior: {
-								X3: [{ Parachain: '1000' }, { PalletInstance: '50' }, { GeneralIndex: '8' }],
-							},
-						},
 						fun: {
-							Fungible: '1000000',
+							Fungible: '100',
 						},
-					},
-					{
 						id: {
-							Parents: '1',
-							Interior: {
-								X3: [{ Parachain: '1000' }, { PalletInstance: '50' }, { GeneralIndex: '1984' }],
+							interior: {
+								Here: '',
 							},
-						},
-						fun: {
-							Fungible: '20000000000',
+							parents: 0,
 						},
 					},
 				],
@@ -352,11 +360,13 @@ describe('ParaToSystem', () => {
 		});
 	});
 	describe('WeightLimit', () => {
+		// NOTE: for V0, V1, and V2 Weightlimit just uses V2 so we only need to test once.
+		// No matter the version if its equal to or less than 2, it will alwyas default to V2.
 		it('Should work when weightLimit option is provided', () => {
 			const refTime = '100000000';
 			const proofSize = '1000';
 
-			const weightLimit = ParaToSystem.createWeightLimit({
+			const weightLimit = RelayToPara.createWeightLimit({
 				weightLimit: {
 					refTime,
 					proofSize,
@@ -364,13 +374,13 @@ describe('ParaToSystem', () => {
 			});
 			expect(weightLimit).toStrictEqual({
 				Limited: {
-					proofSize: '1000',
 					refTime: '100000000',
+					proofSize: '1000',
 				},
 			});
 		});
 		it('Should work when weightLimit option is not provided', () => {
-			const weightLimit = ParaToSystem.createWeightLimit({});
+			const weightLimit = RelayToPara.createWeightLimit({});
 
 			expect(weightLimit).toStrictEqual({
 				Unlimited: null,
