+++ src/createXcmTypes/handlers/SystemToRelay.spec.ts
@@ -1,14 +1,15 @@
 // Copyright 2023 Parity Technologies (UK) Ltd.
 
-import { Registry } from '../registry';
-import { mockMoonriverParachainApi } from '../testHelpers/mockMoonriverParachainApi';
-import { ParaToEthereum } from './ParaToEthereum';
+import { Registry } from '../../registry';
+import { mockSystemApi } from '../../testHelpers/mockSystemApi';
+import { SystemToRelay } from './SystemToRelay';
 
-describe('ParaToEthereum', () => {
-	const registry = new Registry('kusama', {});
+describe('SystemToRelay XcmVersioned Generation', () => {
+	const registry = new Registry('statemine', {});
+
 	describe('Beneficiary', () => {
 		it('Should work for V2', () => {
-			const beneficiary = ParaToEthereum.createBeneficiary(
+			const beneficiary = SystemToRelay.createBeneficiary(
 				'0xf5d5714c084c112843aca74f8c498da06cc5a2d63153b825189baa51043b1f0b',
 				2,
 			);
@@ -30,7 +31,7 @@ describe('ParaToEthereum', () => {
 			expect(beneficiary).toStrictEqual(expectedRes);
 		});
 		it('Should work for V3', () => {
-			const beneficiary = ParaToEthereum.createBeneficiary(
+			const beneficiary = SystemToRelay.createBeneficiary(
 				'0xf5d5714c084c112843aca74f8c498da06cc5a2d63153b825189baa51043b1f0b',
 				3,
 			);
@@ -51,7 +52,7 @@ describe('ParaToEthereum', () => {
 			expect(beneficiary).toStrictEqual(expectedRes);
 		});
 		it('Should work for V4', () => {
-			const beneficiary = ParaToEthereum.createBeneficiary(
+			const beneficiary = SystemToRelay.createBeneficiary(
 				'0xf5d5714c084c112843aca74f8c498da06cc5a2d63153b825189baa51043b1f0b',
 				4,
 			);
@@ -74,7 +75,7 @@ describe('ParaToEthereum', () => {
 			expect(beneficiary).toStrictEqual(expectedRes);
 		});
 		it('Should work for V5', () => {
-			const beneficiary = ParaToEthereum.createBeneficiary(
+			const beneficiary = SystemToRelay.createBeneficiary(
 				'0xf5d5714c084c112843aca74f8c498da06cc5a2d63153b825189baa51043b1f0b',
 				5,
 			);
@@ -99,15 +100,13 @@ describe('ParaToEthereum', () => {
 	});
 	describe('Destination', () => {
 		it('Should work for V2', () => {
-			const destination = ParaToEthereum.createDest('100', 2);
+			const destination = SystemToRelay.createDest('0', 2);
 
 			const expectedRes = {
 				V2: {
 					parents: 1,
 					interior: {
-						X1: {
-							Parachain: '100',
-						},
+						Here: null,
 					},
 				},
 			};
@@ -115,15 +114,13 @@ describe('ParaToEthereum', () => {
 			expect(destination).toStrictEqual(expectedRes);
 		});
 		it('Should work for V3', () => {
-			const destination = ParaToEthereum.createDest('100', 3);
+			const destination = SystemToRelay.createDest('0', 3);
 
 			const expectedRes = {
 				V3: {
 					parents: 1,
 					interior: {
-						X1: {
-							Parachain: '100',
-						},
+						Here: null,
 					},
 				},
 			};
@@ -131,17 +128,13 @@ describe('ParaToEthereum', () => {
 			expect(destination).toStrictEqual(expectedRes);
 		});
 		it('Should work for V4', () => {
-			const destination = ParaToEthereum.createDest('100', 4);
+			const destination = SystemToRelay.createDest('0', 4);
 
 			const expectedRes = {
 				V4: {
 					parents: 1,
 					interior: {
-						X1: [
-							{
-								Parachain: '100',
-							},
-						],
+						Here: null,
 					},
 				},
 			};
@@ -149,17 +142,13 @@ describe('ParaToEthereum', () => {
 			expect(destination).toStrictEqual(expectedRes);
 		});
 		it('Should work for V5', () => {
-			const destination = ParaToEthereum.createDest('100', 5);
+			const destination = SystemToRelay.createDest('0', 5);
 
 			const expectedRes = {
 				V5: {
 					parents: 1,
 					interior: {
-						X1: [
-							{
-								Parachain: '100',
-							},
-						],
+						Here: null,
 					},
 				},
 			};
@@ -168,48 +157,29 @@ describe('ParaToEthereum', () => {
 		});
 	});
 	describe('Assets', () => {
-		const isLiquidTokenTransfer = false;
 		const isForeignAssetsTransfer = false;
+		const isLiquidTokenTransfer = false;
 		it('Should work for V2', async () => {
-			const assets = await ParaToEthereum.createAssets(
-				['1000000000000', '2000000000'],
-				2,
-				'moonriver',
-				['42259045809535163221576417993425387648', '182365888117048807484804376330534607370'],
-				{
-					registry,
-					isForeignAssetsTransfer,
-					isLiquidTokenTransfer,
-					api: mockMoonriverParachainApi,
-				},
-			);
+			const assets = await SystemToRelay.createAssets(['100'], 2, '', [], {
+				registry,
+				isForeignAssetsTransfer,
+				isLiquidTokenTransfer,
+				api: mockSystemApi,
+			});
 
 			const expectedRes = {
 				V2: [
 					{
 						id: {
 							Concrete: {
-								Parents: '1',
-								Interior: {
-									Here: null,
+								parents: 1,
+								interior: {
+									Here: '',
 								},
 							},
 						},
 						fun: {
-							Fungible: '1000000000000',
-						},
-					},
-					{
-						id: {
-							Concrete: {
-								Parents: '1',
-								Interior: {
-									X3: [{ Parachain: '1000' }, { PalletInstance: '50' }, { GeneralIndex: '8' }],
-								},
-							},
-						},
-						fun: {
-							Fungible: '2000000000',
+							Fungible: '100',
 						},
 					},
 				],
@@ -218,45 +188,26 @@ describe('ParaToEthereum', () => {
 			expect(assets).toStrictEqual(expectedRes);
 		});
 		it('Should work for V3', async () => {
-			const assets = await ParaToEthereum.createAssets(
-				['1000000', '20000000000'],
-				3,
-				'moonriver',
-				['182365888117048807484804376330534607370', '311091173110107856861649819128533077277'],
-				{
-					registry,
-					isForeignAssetsTransfer,
-					isLiquidTokenTransfer,
-					api: mockMoonriverParachainApi,
-				},
-			);
+			const assets = await SystemToRelay.createAssets(['100'], 3, '', [], {
+				registry,
+				isForeignAssetsTransfer,
+				isLiquidTokenTransfer,
+				api: mockSystemApi,
+			});
 
 			const expectedRes = {
 				V3: [
 					{
 						id: {
 							Concrete: {
-								Parents: '1',
-								Interior: {
-									X3: [{ Parachain: '1000' }, { PalletInstance: '50' }, { GeneralIndex: '8' }],
+								parents: 1,
+								interior: {
+									Here: '',
 								},
 							},
 						},
 						fun: {
-							Fungible: '1000000',
-						},
-					},
-					{
-						id: {
-							Concrete: {
-								Parents: '1',
-								Interior: {
-									X3: [{ Parachain: '1000' }, { PalletInstance: '50' }, { GeneralIndex: '1984' }],
-								},
-							},
-						},
-						fun: {
-							Fungible: '20000000000',
+							Fungible: '100',
 						},
 					},
 				],
@@ -265,41 +216,24 @@ describe('ParaToEthereum', () => {
 			expect(assets).toStrictEqual(expectedRes);
 		});
 		it('Should work for V4', async () => {
-			const assets = await ParaToEthereum.createAssets(
-				['1000000', '20000000000'],
-				4,
-				'moonriver',
-				['182365888117048807484804376330534607370', '311091173110107856861649819128533077277'],
-				{
-					registry,
-					isForeignAssetsTransfer,
-					isLiquidTokenTransfer,
-					api: mockMoonriverParachainApi,
-				},
-			);
+			const assets = await SystemToRelay.createAssets(['100'], 4, '', [], {
+				registry,
+				isForeignAssetsTransfer,
+				isLiquidTokenTransfer,
+				api: mockSystemApi,
+			});
 
 			const expectedRes = {
 				V4: [
 					{
 						id: {
-							Parents: '1',
-							Interior: {
-								X3: [{ Parachain: '1000' }, { PalletInstance: '50' }, { GeneralIndex: '8' }],
+							parents: 1,
+							interior: {
+								Here: '',
 							},
 						},
 						fun: {
-							Fungible: '1000000',
-						},
-					},
-					{
-						id: {
-							Parents: '1',
-							Interior: {
-								X3: [{ Parachain: '1000' }, { PalletInstance: '50' }, { GeneralIndex: '1984' }],
-							},
-						},
-						fun: {
-							Fungible: '20000000000',
+							Fungible: '100',
 						},
 					},
 				],
@@ -308,41 +242,24 @@ describe('ParaToEthereum', () => {
 			expect(assets).toStrictEqual(expectedRes);
 		});
 		it('Should work for V5', async () => {
-			const assets = await ParaToEthereum.createAssets(
-				['1000000', '20000000000'],
-				5,
-				'moonriver',
-				['182365888117048807484804376330534607370', '311091173110107856861649819128533077277'],
-				{
-					registry,
-					isForeignAssetsTransfer,
-					isLiquidTokenTransfer,
-					api: mockMoonriverParachainApi,
-				},
-			);
+			const assets = await SystemToRelay.createAssets(['100'], 5, '', [], {
+				registry,
+				isForeignAssetsTransfer,
+				isLiquidTokenTransfer,
+				api: mockSystemApi,
+			});
 
 			const expectedRes = {
 				V5: [
 					{
 						id: {
-							Parents: '1',
-							Interior: {
-								X3: [{ Parachain: '1000' }, { PalletInstance: '50' }, { GeneralIndex: '8' }],
+							parents: 1,
+							interior: {
+								Here: '',
 							},
 						},
 						fun: {
-							Fungible: '1000000',
-						},
-					},
-					{
-						id: {
-							Parents: '1',
-							Interior: {
-								X3: [{ Parachain: '1000' }, { PalletInstance: '50' }, { GeneralIndex: '1984' }],
-							},
-						},
-						fun: {
-							Fungible: '20000000000',
+							Fungible: '100',
 						},
 					},
 				],
@@ -356,7 +273,7 @@ describe('ParaToEthereum', () => {
 			const refTime = '100000000';
 			const proofSize = '1000';
 
-			const weightLimit = ParaToEthereum.createWeightLimit({
+			const weightLimit = SystemToRelay.createWeightLimit({
 				weightLimit: {
 					refTime,
 					proofSize,
@@ -364,13 +281,13 @@ describe('ParaToEthereum', () => {
 			});
 			expect(weightLimit).toStrictEqual({
 				Limited: {
-					proofSize: '1000',
 					refTime: '100000000',
+					proofSize: '1000',
 				},
 			});
 		});
 		it('Should work when weightLimit option is not provided', () => {
-			const weightLimit = ParaToEthereum.createWeightLimit({});
+			const weightLimit = SystemToRelay.createWeightLimit({});
 
 			expect(weightLimit).toStrictEqual({
 				Unlimited: null,
