+++ src/createXcmTypes/handlers/SystemToSystem.spec.ts
@@ -1,15 +1,15 @@
 // Copyright 2023 Parity Technologies (UK) Ltd.
 
-import { Registry } from '../registry';
-import { mockSystemApi } from '../testHelpers/mockSystemApi';
-import { SystemToRelay } from './SystemToRelay';
+import { Registry } from '../../registry';
+import { mockSystemApi } from '../../testHelpers/mockSystemApi';
+import { SystemToSystem } from './SystemToSystem';
 
-describe('SystemToRelay XcmVersioned Generation', () => {
+describe('SystemToSystem XcmVersioned Generation', () => {
 	const registry = new Registry('statemine', {});
 
 	describe('Beneficiary', () => {
 		it('Should work for V2', () => {
-			const beneficiary = SystemToRelay.createBeneficiary(
+			const beneficiary = SystemToSystem.createBeneficiary(
 				'0xf5d5714c084c112843aca74f8c498da06cc5a2d63153b825189baa51043b1f0b',
 				2,
 			);
@@ -31,7 +31,7 @@ describe('SystemToRelay XcmVersioned Generation', () => {
 			expect(beneficiary).toStrictEqual(expectedRes);
 		});
 		it('Should work for V3', () => {
-			const beneficiary = SystemToRelay.createBeneficiary(
+			const beneficiary = SystemToSystem.createBeneficiary(
 				'0xf5d5714c084c112843aca74f8c498da06cc5a2d63153b825189baa51043b1f0b',
 				3,
 			);
@@ -52,7 +52,7 @@ describe('SystemToRelay XcmVersioned Generation', () => {
 			expect(beneficiary).toStrictEqual(expectedRes);
 		});
 		it('Should work for V4', () => {
-			const beneficiary = SystemToRelay.createBeneficiary(
+			const beneficiary = SystemToSystem.createBeneficiary(
 				'0xf5d5714c084c112843aca74f8c498da06cc5a2d63153b825189baa51043b1f0b',
 				4,
 			);
@@ -75,7 +75,7 @@ describe('SystemToRelay XcmVersioned Generation', () => {
 			expect(beneficiary).toStrictEqual(expectedRes);
 		});
 		it('Should work for V5', () => {
-			const beneficiary = SystemToRelay.createBeneficiary(
+			const beneficiary = SystemToSystem.createBeneficiary(
 				'0xf5d5714c084c112843aca74f8c498da06cc5a2d63153b825189baa51043b1f0b',
 				5,
 			);
@@ -100,13 +100,15 @@ describe('SystemToRelay XcmVersioned Generation', () => {
 	});
 	describe('Destination', () => {
 		it('Should work for V2', () => {
-			const destination = SystemToRelay.createDest('0', 2);
+			const destination = SystemToSystem.createDest('1000', 2);
 
 			const expectedRes = {
 				V2: {
 					parents: 1,
 					interior: {
-						Here: null,
+						X1: {
+							Parachain: '1000',
+						},
 					},
 				},
 			};
@@ -114,13 +116,15 @@ describe('SystemToRelay XcmVersioned Generation', () => {
 			expect(destination).toStrictEqual(expectedRes);
 		});
 		it('Should work for V3', () => {
-			const destination = SystemToRelay.createDest('0', 3);
+			const destination = SystemToSystem.createDest('1002', 3);
 
 			const expectedRes = {
 				V3: {
 					parents: 1,
 					interior: {
-						Here: null,
+						X1: {
+							Parachain: '1002',
+						},
 					},
 				},
 			};
@@ -128,13 +132,17 @@ describe('SystemToRelay XcmVersioned Generation', () => {
 			expect(destination).toStrictEqual(expectedRes);
 		});
 		it('Should work for V4', () => {
-			const destination = SystemToRelay.createDest('0', 4);
+			const destination = SystemToSystem.createDest('1002', 4);
 
 			const expectedRes = {
 				V4: {
 					parents: 1,
 					interior: {
-						Here: null,
+						X1: [
+							{
+								Parachain: '1002',
+							},
+						],
 					},
 				},
 			};
@@ -142,13 +150,17 @@ describe('SystemToRelay XcmVersioned Generation', () => {
 			expect(destination).toStrictEqual(expectedRes);
 		});
 		it('Should work for V5', () => {
-			const destination = SystemToRelay.createDest('0', 5);
+			const destination = SystemToSystem.createDest('1002', 5);
 
 			const expectedRes = {
 				V5: {
 					parents: 1,
 					interior: {
-						Here: null,
+						X1: [
+							{
+								Parachain: '1002',
+							},
+						],
 					},
 				},
 			};
@@ -156,11 +168,13 @@ describe('SystemToRelay XcmVersioned Generation', () => {
 			expect(destination).toStrictEqual(expectedRes);
 		});
 	});
+
 	describe('Assets', () => {
 		const isForeignAssetsTransfer = false;
 		const isLiquidTokenTransfer = false;
+
 		it('Should work for V2', async () => {
-			const assets = await SystemToRelay.createAssets(['100'], 2, '', [], {
+			const assets = await SystemToSystem.createAssets(['100'], 2, 'statemine', ['USDT'], {
 				registry,
 				isForeignAssetsTransfer,
 				isLiquidTokenTransfer,
@@ -172,9 +186,9 @@ describe('SystemToRelay XcmVersioned Generation', () => {
 					{
 						id: {
 							Concrete: {
-								parents: 1,
-								interior: {
-									Here: '',
+								Parents: '0',
+								Interior: {
+									X2: [{ PalletInstance: '50' }, { GeneralIndex: '11' }],
 								},
 							},
 						},
@@ -188,7 +202,7 @@ describe('SystemToRelay XcmVersioned Generation', () => {
 			expect(assets).toStrictEqual(expectedRes);
 		});
 		it('Should work for V3', async () => {
-			const assets = await SystemToRelay.createAssets(['100'], 3, '', [], {
+			const assets = await SystemToSystem.createAssets(['100'], 3, 'bridge-hub-kusama', ['ksm'], {
 				registry,
 				isForeignAssetsTransfer,
 				isLiquidTokenTransfer,
@@ -200,8 +214,8 @@ describe('SystemToRelay XcmVersioned Generation', () => {
 					{
 						id: {
 							Concrete: {
-								parents: 1,
-								interior: {
+								Parents: '1',
+								Interior: {
 									Here: '',
 								},
 							},
@@ -216,7 +230,7 @@ describe('SystemToRelay XcmVersioned Generation', () => {
 			expect(assets).toStrictEqual(expectedRes);
 		});
 		it('Should work for V4', async () => {
-			const assets = await SystemToRelay.createAssets(['100'], 4, '', [], {
+			const assets = await SystemToSystem.createAssets(['100'], 4, 'bridge-hub-kusama', ['ksm'], {
 				registry,
 				isForeignAssetsTransfer,
 				isLiquidTokenTransfer,
@@ -227,8 +241,8 @@ describe('SystemToRelay XcmVersioned Generation', () => {
 				V4: [
 					{
 						id: {
-							parents: 1,
-							interior: {
+							Parents: '1',
+							Interior: {
 								Here: '',
 							},
 						},
@@ -242,7 +256,7 @@ describe('SystemToRelay XcmVersioned Generation', () => {
 			expect(assets).toStrictEqual(expectedRes);
 		});
 		it('Should work for V5', async () => {
-			const assets = await SystemToRelay.createAssets(['100'], 5, '', [], {
+			const assets = await SystemToSystem.createAssets(['100'], 5, 'bridge-hub-kusama', ['ksm'], {
 				registry,
 				isForeignAssetsTransfer,
 				isLiquidTokenTransfer,
@@ -253,9 +267,50 @@ describe('SystemToRelay XcmVersioned Generation', () => {
 				V5: [
 					{
 						id: {
-							parents: 1,
-							interior: {
+							Parents: '1',
+							Interior: {
 								Here: '',
+							},
+						},
+						fun: {
+							Fungible: '100',
+						},
+					},
+				],
+			};
+
+			expect(assets).toStrictEqual(expectedRes);
+		});
+
+		it('Should error when asset ID is not found for V3', async () => {
+			const expectedErrorMessage = 'bridge-hub-kusama has no associated token symbol usdc';
+
+			await expect(async () => {
+				await SystemToSystem.createAssets(['100'], 3, 'bridge-hub-kusama', ['usdc'], {
+					registry,
+					isForeignAssetsTransfer,
+					isLiquidTokenTransfer,
+					api: mockSystemApi,
+				});
+			}).rejects.toThrow(expectedErrorMessage);
+		});
+		it('Should work for a liquid token transfer', async () => {
+			const assets = await SystemToSystem.createAssets(['100'], 2, 'statemine', ['USDT'], {
+				registry,
+				isForeignAssetsTransfer,
+				isLiquidTokenTransfer: true,
+				api: mockSystemApi,
+			});
+
+			const expectedRes = {
+				V2: [
+					{
+						id: {
+							Concrete: {
+								Parents: '0',
+								Interior: {
+									X2: [{ PalletInstance: '55' }, { GeneralIndex: '11' }],
+								},
 							},
 						},
 						fun: {
@@ -273,7 +328,7 @@ describe('SystemToRelay XcmVersioned Generation', () => {
 			const refTime = '100000000';
 			const proofSize = '1000';
 
-			const weightLimit = SystemToRelay.createWeightLimit({
+			const weightLimit = SystemToSystem.createWeightLimit({
 				weightLimit: {
 					refTime,
 					proofSize,
@@ -287,7 +342,7 @@ describe('SystemToRelay XcmVersioned Generation', () => {
 			});
 		});
 		it('Should work when weightLimit option is not provided', () => {
-			const weightLimit = SystemToRelay.createWeightLimit({});
+			const weightLimit = SystemToSystem.createWeightLimit({});
 
 			expect(weightLimit).toStrictEqual({
 				Unlimited: null,
